Move Insert, Update and Delete functions to a package
