--- go.mod
@@ -1,3 +1,0 @@
-module github.com/bend-is/avto-dev-go-lang-lessons
-
-go 1.15
