+++ task2/go.mod
@@ -1,0 +1,3 @@
+module "task2"
+
+go 1.15
