+++ task2/pkg/sortedarray/slice.go
@@ -1,25 +1,7 @@
-package main
+package sortedarray
 
-import "fmt"
-
-func main() {
-	var numbers []int
-	var inputNumber int
-
-	for {
-		if _, err := fmt.Scan(&inputNumber); err != nil {
-			break
-		}
-
-		numbers = UpdateNumbers(numbers, inputNumber)
-
-		// Clear the output and print current numbers
-		fmt.Printf("\u001B[H\u001B[2J\n%v\n", numbers)
-	}
-}
-
-//UpdateNumbers update slice of numbers. If new number is positive it will be inserted else deleted from slice.
-func UpdateNumbers(numbers []int, number int) []int {
+//Update update slice of numbers. If new number is positive it will be inserted else deleted from slice.
+func Update(numbers []int, number int) []int {
 	if number > 0 {
 		numbers = InsertV2(numbers, number)
 	} else {
